[access.ffuse]: Add read_only option

+++ ufs/access/ffuse.py
@@ -13,23 +13,25 @@ from ufs.utils.pathlib import SafePurePosixPath
 from ufs.access.shutil import walk, copytree, rmtree
 
 @contextlib.contextmanager
-def ffuse_mount(ufs: UFS, mount_dir: str = None):
+def ffuse_mount(ufs: UFS, mount_dir: str = None, readonly: bool = False):
   mount_dir_resolved = pathlib.Path(mount_dir or tempfile.mkdtemp())
   mount_dir_ufs = Prefix(Local(), mount_dir_resolved)
   root = SafePurePosixPath()
   copytree(ufs, root, mount_dir_ufs, root, exists_ok=True)
-  before = OrderedDict(walk(mount_dir_ufs, root, dirfirst=False))
+  if not readonly:
+    before = OrderedDict(walk(mount_dir_ufs, root, dirfirst=False))
   try:
     yield mount_dir_resolved
   finally:
-    after = dict(walk(mount_dir_ufs, root, dirfirst=False))
-    for path in (before.keys() - after.keys()):
-      item = before[path]
-      if item['type'] == 'file':
-        ufs.unlink(path)
-      elif item['type'] == 'directory':
-        ufs.rmdir(path)
-    copytree(mount_dir_ufs, root, ufs, root, exists_ok=True)
+    if not readonly:
+      after = dict(walk(mount_dir_ufs, root, dirfirst=False))
+      for path in (before.keys() - after.keys()):
+        item = before[path]
+        if item['type'] == 'file':
+          ufs.unlink(path)
+        elif item['type'] == 'directory':
+          ufs.rmdir(path)
+      copytree(mount_dir_ufs, root, ufs, root, exists_ok=True)
     rmtree(mount_dir_ufs, root)
 
 if __name__ == '__main__':
@@ -37,5 +39,5 @@ if __name__ == '__main__':
   ufs = UFS.from_dict(**json.loads(os.environ.pop('UFS_SPEC')))
   mount_dir = pathlib.Path(sys.argv[1])
   assert mount_dir.exists()
-  with ffuse_mount(ufs, mount_dir):
+  with ffuse_mount(ufs, mount_dir, bool(os.environ.pop('UFS_READONLY', ''))):
     threading.Event().wait()
